Fixed ts issues
