+++ src/app/app/(common)/account/page.tsx
@@ -17,7 +17,7 @@ const Account = () => {
       
       <div className="grid grid-cols-1 md:grid-cols-3 gap-6">
         <div className="md:col-span-2">
-          <UserProfile user={user} />
+          <UserProfile user={user || undefined} />
           <RecentActivity />
         </div>
